requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
openai>=1.0.0
ollama>=0.1.0
tabulate>=0.9.0
openpyxl>=3.1.0
